refactor: rename directory naming Blog to Post
