+++ common/routes/index.ts
@@ -1,5 +1,5 @@
 type NavLink = { title: string; link: string };
 
-const navLinks: NavLink[] = [{ title: "Blog", link: "/blog" }];
+const navLinks: NavLink[] = [{ title: "Posts", link: "/posts" }];
 
 export default navLinks;
